Update patch set 1
Patch Set 1:

Test case: Start a search with "a" as the keyword, which
will cause the "search terms has a small number of characters, do you want to proceed?" dialog
to come up. Here, clicking "Yes" will actually cancel the operation, but clicking "No" would allow it to continue.

Patch-set: 1