Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

New landscape layout mode (onConfigurationChanged not allow to use layout-land, so in every onConfigurationChanged, the layout is adjusted to fit the orientation).

The status bar with the buttons is now in the top title bar in landscape mode.

Patch-set: 5