Update patch set 1
Patch Set 1:

Engle, great catch.

r.filter is null because the interanl editor is not part of the queryIntentActivities search. In normal conditions (from a query activity search) the r.filter will not be null (this part of the code is copied from the aosp dialog).
Reviewing the code i prefer do the next:

move !isInternalEditor(ri) check from

    if (isPlatformSigned && this.mAllowPreferred) {
	if (filter != null && !isPreferredSelected() && !isInternalEditor(ri)) {
	    try {

to the first line of the method

        if (remember) {

so, if the internal editor is selected then nothing is do about the resolve of the intent for  system association (the internal editor mustn't be associated in any case).

What do you think? Pull this changes if you are conform with it.

Patch-set: 1