Update patch set 1
Patch Set 1:

David, the dark theme is in the CMFileManagerThemes apk, that is under theme folder. You need to build and install it. After you install it, the theme will be listed, as well as others their party apks with themes designed for the file manager apk.

If you are using the cm build system, enable the compilation and distribution of the theme apk adding the theme project to vendor/cm/config/common.mk por vendor/cm/config/themes_common.mk

Patch-set: 1