Update patch set 3
Patch Set 3:

David, when you work with all the android code under eclipse there is not need for a gitignore because none of this files is created by eclipse under the project.
When i created the project i worked separately of cm code and under a the ADT project. In this case all of cm non-relevant files created by eclipse must be ignored, for avoid accidental uploads, and at least one gitignore (at root project) sould be uploaded. In this case i prefer to upload 3 files (one per every project).

Patch-set: 3