Update patch set 1
Patch Set 1:

Yes, all cyanogenmod are rooted. CMFM has a special mode when device is not rooted. Instead of use the complete shell console, it used a java based console with limited operations. This change is more for stocks roms. To enable it, you can force the value of sIsDeviceRooted to false in register method of FileManagerApplication class.

Patch-set: 1