Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This discussion is entirely besides the point of this patch though, isn't it?

MIME type resolution for downloads has failed on me a number of times already, which is where I truly welcomed the file manager. So far I needed the root access only for (rare) debugging purposes, but as it's not enabled-by-default behaviour, I don't see a problem in having it there.

Patch-set: 1