Update patch set 1
Patch Set 1: Looks good to me, approved

It's because this commit removes all the commments (new policy) and changes a tab to consist of four instead of two spaces. There are no new translations/translation changes, so I'll merge this one right now.

I'll create a new change that will add the missing translations (also in the right order).

Patch-set: 1