Update patch set 1
Patch Set 1:

@Jorge, I understand what you mean, but a Paste action is also done when Moving, so it is confusing.

MOVING a file means to first CUT the file (not copy) and then Paste it

COPYING a file means to first COPY the file and then Paste it.

Since they both have the Paste action in them, the menu buttons should be named after the action that you are really doing, i.e. MOVE and COPY

Patch-set: 1