Update patch set 1
Patch Set 1:

As far as I can see it's not done normally and it's up to translators to remove unused strings. Translation Tool used by most of CyanogenMod translators handles it as well. And as a translator I'd hate if it would be done - it would create a merge conflicts with uncommitted changes.

Patch-set: 1