Update patch set 2
Patch Set 2:

I think that's because there is no need for in English. When the quantity is 0, the item for quantity 'other' is used and that generates a correct sentence. However, some languages might have grammar rules for different quantities, but I don't think that the author does not have the knowledge to decide to remove or to retain that item in each supported language :)

It's different from how (string) arrays are handled.

Patch-set: 2