Update patch set 10
Patch Set 10: Code-Review+1

I wouldnt know what kind of icon you would give music that is dynamic a wave which is kindda useless anyway?

I tested this works very good for pics and apps even the async part.

Something to note which might not be clear to users is that if your are browsing a folder that needs root access thumbs will not be generated (cause the java code used to make thumbnails has no access to the files i assume) which for me isnt really a dealbreaker.

Patch-set: 10