Update patch set 10
Patch Set 10:

Jo De Boeck, i'm thinking on album covers for folders (but i think we have to relay on an specific music app. AFAIK is not present in media database). Anyway, just a thought.

With the exception of /system/media/video, media data (pictures and videos) musn't be on system partitions (apps works because they are extracted from android database).

Patch-set: 10