Update patch set 1
Patch Set 1:

Making assumptions as to where the real mountpoint is is dangerous. Google has been shifting this around with almost every release, and I doubt they're done; I suspect the message they're trying to send is "use the damn API to figure out where storage is instead of assuming" :-)
We can't use the API to figure the real paths, but there are variables with that data. As system components, we should get this from the environment if we can (regardless of doing this here or under the superuser bits)

Patch-set: 1