Update patch set 1
Patch Set 1:

I guess my rationale behind this is that most people will look to a folder such as /sdcard or /extSdCard when looking for their internal/external sdcards.

Its a bit clumsy to figure out how to turn on symlinks if you don't know where to do it.

My apologies if this has been a discussed topic before and I missed it. In such a case, review and merge it anyway =P

Patch-set: 1