Update patch set 13
Patch Set 13: Code-Review-1

The media file thumb generation doesn't work in root mode (as mentioned by Jo above), which I'm pretty sure happens because of the /storage/emulated vs. /mnt/shell/emulated mismatch. So we should either handle this problem for the thumb generation as well or fix the su daemon to provide the respective mount instead.

(And did you deliberately choose not to move the icon cache to the application scope, or did you just miss my respective comment in PS12 and/or just didn't have time yet?)

Patch-set: 13