Update patch set 17
Patch Set 17:

Danny, what i see if that PS17 performs worst than the previous patch. In a folder with 426 images of relative high quality, scrolling when images are not cached is a painful :(

Also, i got an strange dead lock when tapping another folder in the breadcrumb while staying in this folder (i cannot reproduce it any more)

I doing something similar here https://github.com/jruesga/android_packages_wallpapers_PhotoPhase/blob/cm-10.2/src/org/cyanogenmod/wallpapers/photophase/widgets/AlbumPictures.java that performs good, but it not a listview. In my experience with listviews, the getView should do the less work as possible, but it just my feeling :)

Anyway, if you are confortable with it, then see my previous comment about caching the normailzed media path and ship it (this only affects to directories with a lot of images)

Patch-set: 17