Update patch set 4
Patch Set 4:

Yes, the new mount console is only for 'global' mounts. Depending on whether we want to make the mounts done by CMFileManager visible to other processes or not (I'd say it's not strictly needed to have this), we can get away without the mount console if we find a way to apply the remounts to both foreground and background console (basically doing the remount twice).

With my current su code changes done in the global mount namespace don't yet propagate into the children which hava eMOUNT_EXTERNAL_STORAGE set, though. I need to look into that (and am open to any suggestions about that ;) ).

Patch-set: 4