Update patch set 6
Patch Set 6:

Thank you for reviewing!

I'll look into the theme issue and fix the smaller ones the next days as soon as I find time.

The only thing I have to disagree about is the opening of the drawer on the app's first launch. As it says in the official design guidelines (https://developer.android.com/design/patterns/navigation-drawer.html), "introduce the user to the navigation drawer by automatically opening it. This ensures that users know about the navigation drawer".

Patch-set: 6