Update patch set 1
Patch Set 1: Code-Review-1

Please wait with merging this, give people a little more chance to re-add the changed strings to Crowdin.

There weren't really any changes, only a few dots removed, so it would be a shame if a lot of strings got removed because of this.

Patch-set: 1
Label: Code-Review=-1