Update patch set 3
Patch Set 3:

(1 comment)

I believe it would be beneficial to the end user if we display a warning or toast detailing that a secure file system synchronization is in progress, otherwise it may confuse the user who is attempting to access the secure virtual mount point before the synchronization has completed.

Will have more comments soon.

Patch-set: 3