Update patch set 3
Patch Set 3:

Adnan, I'm not agree about warning the user on every sync delayed operation. It's just like fsyc operation in a kernel filesystem. The kernel wasn't advise you when it performs the synchroniation to the underlying filesystem. It simply does it. This is a bit the same, when there isn't user activity (a secure filesystem belongs only to the current user, so the activity alwaya comes from him) over the secure filesystem the synchronization happens in background. The problem here is that the original truezip shutdown hooks doesn't work in Android, and there us a secure way to unmount the filesystem, other than perform a sync operation when a write operation happens. This is done after 20 seconds of inactivity or after the write operation (depending this user setting). Notify the user on every will be annoying ihmo.

About the selection, you are right. I failed here. Secure mountpoint (root folder) never should accept move/copy/rename operations. And I forget to remove paths from the secure filesystem in the current selection if the secure storage was unmounted. I will fixed that tonight.

Thx for the review.

Patch-set: 3