Update patch set 2
Patch Set 2:

(1 comment)

So if i understood correctly, the problem isn't with the Gallery2 itself, but with the fact that we want to return to CMFM when it opened an image mimetype, right? In that case, we should filter by mimetype instead of package. Making rules for specific packages doesn't sound good to me.

Patch-set: 2