Update patch set 2
Patch Set 2:

I think the problem isn't with all image mimtypes. For example, if the user chooses to use Google Photos to respond to this intent, pressing the back button will return the user to CMFM.

Thus, seems the "set back button as up" is specific to the Gallery2 app and was added in AOSP to handle all intents with the flag "NEW_TASK" in this fashion. So a package name filter may still be the most appropriate fix, however, I do agree that an "equals" check is more appropriate than the "contains" check.

Patch-set: 2