Update patch set 1
Patch Set 1:

Offtopic, but I need some help with RTL. In CM we support Kurdish which is not supported in AOSP and is supposed to be RTL, but according to the users it appears as LTR. It was working fine in cm-11.0. Does anyone know if there is a place where RTL languages are defined?

Patch-set: 1