Update patch set 2
Patch Set 2:

(4 comments)

If i understood correctly this bug only can happends in sdcards formatted with ntfs/vfat. That means that the bug shouldn't occurs in a priviledged folder, rigth? I wonder that because we cannot use a java operation if it will require an elevation priviledge. I can think in the obb folder or other folder behind the user space folder.

Maybe instead of use a java command we can do atemporary "double" move using the current console, but not sure there is a good use case that justify this performance penalty for insensitive filesystems.

Patch-set: 2