Support ambigous file extension mimetypes.
Previously, CMFileManager operated under the assumption that file
extensions map to exactly one mimetype. This is not true in some cases,
such as .3gp files, which can have an audio or video mimetype.

Add support so that in mime_types.properties we can specify a comma
separated list of mimetype info that are matched with a given extension.
If an AmbiguousExtensionHelper subclass implementation is provided for one of these
extensions, this is used to determine the correct mimetype for the file.

Change-Id: Ie73d6ad646692dfeac112ac50c1c6436e6b5559b