Update patch set 1
Patch Set 1:

Yeah I think that is only for something using a virtual console vs the shell console.  This patch doesn't involve cleanup yet, and also doesn't account for that I think.  I was unsure about how all of that worked.  It seems if it needs to it will copy it to external for access by all virtual console but I think that may be if you're copying something not on external.

Patch-set: 1