Update patch set 1
Patch Set 1:

oh I see, so it tries to get access to a file and if the access mode is safe and the storage dir is not a directory then you will return a "chrooted virtual storage dir" which is just a file with the base path equal to external storage directory.  What constitutes as a "Virtual Mount Point"?

Patch-set: 1