Update patch set 4
Patch Set 4:

I was trying to test this and I thought your patch wasn't working before I realized the OK button gets grayed out. Can't we just maxLength on the EditText so the user just cannot enter any more text after 255? Feel like that might be more indicative to the user, otherwise we'd really have to set a message on the InputDialog to explain why the user cannot pick this folder/file name as is done in the other error cases.

Patch-set: 4