Update patch set 8
Patch Set 8:

Thanks for pointing that out, Jorge. We would like to support multiple encodings of text in EditorActivity. Currently, the JavaConsole preserves the original encoding when the file is read, but the ShellConsole does not. I think that to solve this I have two strategies I can take:

1. Allow for a PrivilegedJavaConsole that can read files in the root partition, if this is possible. That way, I can still read the files using the JavaConsole.

or

2. I need to change ShellConsole so that the output of the command is not parsed as a string with the system default encoding and I can use the correct encoding.

Do you suggest one strategy over another?

Patch-set: 8