Update patch set 8
Patch Set 8:

Matt, from my knowledge the option 1 isn't possible, because there is no way i know to elevate privileges in a unprivileged proccess (the java one).
About 2, I think it could work, but in the console, it need to decode the text to find the tokens of every command request. That is done in the ShellConsole. You can parse to string to find these tokens but returns a byte array.

Another solutions (that implies more time to code) are:

1.- Create a native executable that can be extracted from the apk, and then su this process, and intercommunicate with that process via RPC or socket (We can create a binder or a namepipe to this proces). The gain with that is that that process can handle multiple request in just one process.

2.- Create a system level api for file system operations (since system runs in with more elevates privileges) and provide access to list, create,... operations via a system service. The gain with that is a public api that can be used by all the file managers apps.

Patch-set: 8