Update patch set 8
Patch Set 8:

Thanks Jorge. I think we are going to try to change ShellConsole to return the original byte array output of /system/bin/cat instead of a String. Those other solutions are very interesting though. I think that in the future we could perhaps implement it that way.

Roman and I are taking a look at this solution today. Nir, we plan to have a new patch here to merge. I only merged a version of this in the other branch because that project doesn't use the ShellConsole, so it was unaffected by this problem. We'll merge this patch when it's ready.

Patch-set: 8