Update patch set 15
Patch Set 15:

I can still get the app to lock up with or without this patch when opening large files. It seems like it might be a little be worse with this addition, but I'm not sure. This is never a problem when using the JavaConsole.

I think we could ship this patch as is and handle the bad behavior in a different patch, since it existed before this patch.

Thoughts?

Patch-set: 15