Update patch set 2
Patch Set 2: Verified-1 Code-Review-1

Still having some weirdness with the bookmarks in the sliding drawer in RTL. I need to compare to the regular layout and figure out where the icons should go in RTL.

In the FSO properties dialog, the permissions strings move each time you open the tab in RTL mode for some reason. That needs to be addressed too.

Patch-set: 2