Create patch set 6
Uploaded patch set 6.

Patch-set: 6
Subject: Implement a dialog that warns the user that we must expose the content of the file by copying it out to an unsecure location in order to allow the external applications to read the files.
Commit: 55186710797becc9a65e2e23e3f04b612c277dbb