Update patch set 5
Patch Set 5:

Martin, doesn't the data served by the SecureResourceProvider provider is via a pipe. I don't see where the RPC limit is a problem here. It can't be resolved witht a LocalSocket rather than a real server socket?

I don't like too much the idea of having a server socket thread in an app (and with privileges to access a secure filesystem). And if iirc we only have a background console, so we can't know handle multiple requests via the http server.

Patch-set: 5
Label: Code-Review=0