Update patch set 1
Patch Set 1:

I know this is probably a nitpicky thing since we don't have a spec for the length, but if the limit was intended to be 255 as defined in InputNameDialog.FILENAME_CHAR_LIMIT, won't this break that contract? Now the user can enter 256 characters and then will get the message after they enter the 256th character?

Wouldn't the accurate solution be to change line 265 of InputNameDialog.java to >= instead of > ?

Patch-set: 1