Update patch set 2
Patch Set 2:

In most cases,there is only one but if some times ,there is more than 1.such as there is a process com.android.musicand there are two activities can open music file via CMFileManager,this is com.android.music.Music and com.android.music.PreviewMusic.
and the priority the PreviewMusic is higher than Music.We open a music file in CMFileManager and select Music remember ,the next time open a musicfile ,it always open by PreviewMusic not Music.I think it's a bug.
    If the size more than one and there don't find right activity.I think the return null and make user choose again is a right approach.

Patch-set: 2