Update patch set 1
Patch Set 1:

That sounds beyond the scope of File Manager. Shouldn't it simply allow the user to open a file with a given mimetype? What if some proprietary app that they have installed can handle DRM protected m4v files? In that case, this is useful to the user as that app would respond when they open it.

Patch-set: 1