Update patch set 1
Patch Set 1: Code-Review+1

I think the counter example case is one in which a user has an m4v file with DRM that they want to play and a player that can play it successfully but ONLY responds to the x-m4v mimetype. With this patch that app will not respond to the intent filter for the DRM m4v file correctly.

However, it appears that this is really only a problem for iTunes and this is not standardized. The standard mimetype is the video/mp4, like you have changed it to you. Source: http://stackoverflow.com/a/15279480/815679

After reading that SO answer, I say +1.

Patch-set: 1