Update patch set 1
Patch Set 1:

It because it's an Itunes DRM encoded file, It won't be opened anywhere besides Itunes unless stripping the DRM protection. However, the likelihood of files like this being played by android users is really low, and much more likely is to play normal, non-DRM m4v file, which this change is needed for, as stated above by Matt. In the end, this change has more benefits than leaving it as is right now.

Patch-set: 1