Update patch set 2
Patch Set 2:

Think the change makes sense, I still need to test it with easy mode on.

Just wanted to note that it's interesting CMFM will display a toast saying "press back again to exit" and the system shows a toast after saying that the screen is pinned. Not sure what we can do to fix this, but I wanted to throw it out there in case any one has any good ideas to fix that toast from showing up in screen pinned mode.

Patch-set: 2