Update patch set 5
Patch Set 5:

deleted the jpg and commented the part in navigation.xml locally to have it compile and see how it looks. Icons seem to be too big and everything is dark grey looks like there should be color and is missing (i know it is intentional). I don't like the teal color in header as well, usually have that color in more decent places only

Patch-set: 5