Invalidate preferences headers when resuming
If the use is in a sub settings, i.e. general settings, and changes
their language.  Upon returning to File Manager and pressing back
to go to the main settings menu, the about text will remain in the
previous language.  This is easily reproduced when the intial language
is set to Chinese and then switched to English.  The about text will
still be in Chinese.

Change-Id: I1dff61a6a94e6aabf336d863e102c29044f3a649