Update patch set 1
Patch Set 1:

This is a valid problem, but I don't think this is the proper solution.

When you write the bytes into data[], you are using the stdin.toString().substring, but that forces the string into UTF-8 (or system default) encoding, which in turn, later breaks trying to read files not encoded in UTF-8.

What we are really trying to do here is match our start and end patterns, which we define, so we can assume it's *always* UTF-8, and just compare the raw bytes instead of trying to use the pattern matcher.

Patch-set: 1