Update patch set 2
Patch Set 2: Code-Review-1

I'm not seeing a change in behavior with this patch.  When in portrait I activate the sort popup menu and then rotate to landscape.  The popup remains but cannot be scrolled until I dismiss it and click it again.

Going the opposite way, from landscape to portrait, the popup remains the height from landscape with scrolling of the content.

Patch-set: 2