Update patch set 8
Patch Set 8:

@Martin

I've addressed the issue you raised, if the buffer there fails to read out, the entire operation should fail (like in other exceptions).

@All

I've updated the logic concerning when to try to load a file URI vs when to load a Content URI. For some reason, content URI's would return a path with our method and cause the editor to try to load random folders in the data partition. Now we check the schema of the URI beforehand and capture.

Patch-set: 8