Fix selection state when browsing storage provider
When browsing storage provider, if an item is selected and you browse to
a new directory which happens to also have a file of the same name, that
file is also show as selected. However, only one selection is shown in
the snackbar.
This patch fixes the FileSystemObject equals method, and fixes the
createFileSytemObject method for creating storage provider FSOs.

Change-Id: Ia69818c4d1b2b580cd881ea25a44f2b9e6aea3b7