Update patch set 1
Patch Set 1:

I think this is a roundabout way if solving a greater issue (Assuming you are just trying to fix the jump/remove the flicker).  The problem at the moment is a refresh (replace of data) causes the data set to jump. A refresh of a dataset should never cause a jump like this. I would explore why this is happening.

Patch-set: 1