Update patch set 1
Patch Set 1: Code-Review+2

Code looks fine, using the copy mechanic seems to make sense for now, but we can't show the src file in it's parent directory while this move is pending or in progress. This doesn't need to be solved until we take on the new copy/move UI, but something to keep an eye on.

Seems to work for me as far as I could tell. Waiting to see if Ed has any feedback tomorrow morning before merging

Patch-set: 1