Fix Secure SD Card Entry to open folder.
Whenever the secure folder was opened from the drawer, it would prompt
for a password and then go to the root folder again. This was caused
by the logic in onResume() which is all meant for when entering
another activity and then coming back into this activity. This should
really be the job on onStart() as onResume() also handles dialogs.

Change-Id: I45e0e44e3e8f9031b0685628a3f24c02725e24e1
Ticket: QRDL-946
(cherry picked from commit 9426b243dd41dea57718efa922e9c5f345338472)