Update patch set 1
Patch Set 1:

But some of the logic of onResume is still valid though. If one resume the activity, and externally remove the directory, i guess we still want to check the current directory and move to default directory to avoid bad operations. onStart is not called if the activity is only resumed.

Patch-set: 1