Update patch set 1
Patch Set 1:

I forgot to address the lifecycle question -- the only time onResume() will be called without onStart() is when a new Window within the activity is opened.  An example of this would be a dialog.  Again, the receivers should be handling this case as they are registered/unregistered in onCreate()/onDestroy()

Patch-set: 1