Update patch set 1
Patch Set 1:

> Jon, sorry my bad. forgot that onStart is called on activity
 > resume.

Actually, I think this whole situation would be solved by using a FileObserver on the current folder instead of doing what we're doing -- This would actually handle quite a bit more cases such as removing folders or items in a folder via ADB, etc. and keep the list up to date as things change external to the app.

Also, as a sanity check, I went back over the docs

Patch-set: 1