Made NavigationView reliant on FileObserver to keep up to date.
We no longer deal with having to refresh a directory upon reentering.
Instead we now just listen on the directory for changes and update as
necessary.  This will make it much more difficult to get out of sync
with the file system in pretty much all situations.

Changed per comments on patch to no longer do refreshes for the list
when items change as the FileObserver handles all of this.

Ticket: QRDL-932
(cherry picked from commit 7b09bb0f1390b72d0ac18484721998eaf805e605)

Conflicts:
	src/com/cyanogenmod/filemanager/activities/NavigationActivity.java

Change-Id: I1c2c4a852e61c6549d87ae07a03040d57f1fffdd