Made NavigationView reliant on FileObserver to keep up to date.
We no longer deal with having to refresh a directory upon reentering.
Instead we now just listen on the directory for changes and update as
necessary.  This will make it much more difficult to get out of sync
with the file system in pretty much all situations.

Changed per comments on patch to no longer do refreshes for the list
when items change as the FileObserver handles all of this.

Change-Id: I28a6af965708ce76107a09d6b931f4c7bafb8365
Ticket: QRDL-932
(cherry picked from commit 7b09bb0f1390b72d0ac18484721998eaf805e605)