ANR Caused by passes over Data Set on notifyDataSetChanged()
Logic existed which took a pass over the data on the UI thread each
time notifyDataSetChange was called which can cause an ANR on a large
data set. This fix removes a lot of the unnecessary logic associated
with this and moves it to getView(). Performance still seems good on
the OnePlus for scrolling.

Change-Id: Idde74f0688fef0ba7e4b560b06d9494896a24174
Ticket: QRDL-931