Update patch set 3
Patch Set 3:

If you are referencing the activity from the weak reference in doInBackground(), you either need to instead use the Application context or reference the particular thing you need which doesn't hold a reference to the Activity context. I haven't looked at it super-carefully, so I don't know if there is even a solution for this given the current code base.

Patch-set: 3