Fixed Misunderstanding of what the FileObserver can listen on
I had a misunderstanding about what the FileObserver could listen on
and broke stuff associated with Root privileged as well as Secure
storage refreshing.  This fix basically changes how we deal with
FileObservers and put them more in tune with the actual FS notifying
us of changes when we can. On a per-console basis, we now listen for
changes to the file system, and it is up to the console to implement
when to notify the listeners of changes.  This removes notification
logic from having to happen in multiple locations and instead
comes straight from the source more similar to dealing with
ContentProviders.

One caveate is that the Shell Console still acts in the old manner
of having to refresh upon coming back from another activity.  On top
of this, we only get notifications for changes made internal to the
app when we are not using the JavaConsole.

Change-Id: Id87b15348370553ba2df5ef92c71b2d993106626
Ticket: QRDL-972