Fixed ANR associated with SecureStorage
An ANR was happening on SecureConsole when a very large item was added
and then you go and try and open the menu options for a SecureConsole
item.  This is caused by the locking mechanism used by SecureConsole.
Consoles, long-term, need to be refactored to not synchronize on the
SecureConsole object itself.

Change-Id: I845372567b8656d63192bfde27952240915ecfe6
Ticket: QRDL-971