Fixed ANR associated with SecureStorage
An ANR was happening on SecureConsole when a very large item was
added, canceled, and click the parent in the Secure folder. This is
caused by the locking mechanism used by SecureConsole. Consoles,
long-term, need to be refactored to not synchronize on the
SecureConsole object itself during execute. The only fix we can do
without full refactoring is to not allow cancelling the operation
because the Console doesn't support the cancel operation.

Change-Id: I845372567b8656d63192bfde27952240915ecfe6
Ticket: QRDL-971