ListPopupWindow: set height of the list
We need to make sure that the list is the proper height
so that it does not draw off the screen.

Ideally, we would stop handling the orientation changes and allow
the app to do the orientation changes as android apps do normally.
This would require some significant changes to how our dialog
system works and could cause other potential issues.

Ticket: QRDL-882
Change-Id: I39176d56d8286ed93b71dbe4c770fe79f39ac753