Fixed ANR & Memory Leak Associated with 3GP
We are now defaulting to treating 3GP files as video files for
determining the default icon as it was parsing the video files for
their metadata which was causing ANR problems. On top of this, it was
discovered that the metadata parser was leaking memory.

Change-Id: I88f6cf3d8ae1a62d1294bd3272b27715c2352525
Ticket: QRDL-931
(cherry picked from commit 6f6094da21960569b93f6171e39e211e8d37f41a)