Update patch set 2
Patch Set 2: Code-Review-1

The long clickable is set in the onTouch listener -- If you're going to handle it in the listener, I'd also handle setting it in the listener for when the mOnItemFlingerListener exists as well and remove it from onTouchEvent (include setClickable() as well)-- On top of this, if you really want to fix it, make it so we can do touch, long press, and swipe all at the same time. It's a little silly we can't support all of them.

Patch-set: 2