Check If System Wants Intent
After much thought about this last night, I came to the conclusion
that because applications can in fact do this, we need to be able
to support it in a backwards compatible fashion. In doing so, there
are a few cases which won't work with suggestions, however, the
changes here are less severe than what they were before. Now just
an extra check happens to see if the system wants to handle the
intent if nothing was found to handle it.

What this means is any application acting as its own resolver
will not automatically have the system handle it in this case
and we no longer can support suggestion outlined here:
http://developer.android.com/training/basics/intents/sending.html#StartActivity

The suggestion is wrong anyways as an activity isn't guaranteed
to continue to exist after the query when the start is called. Maybe
I'll try and push a documentation change upstream to AOSP for this.

Change-Id: I8510660420ee52c09d03c719850fa14d6b1c4441
Issue-Id: CYNGNOS-1152
(cherry picked from commit d28d1480987de805421689831593619e653e659e)