Update patch set 3
Patch Set 3:

But I was design that way. There is no way to open a binary file, because no app can open it. But the internal editor can open it as a hexdump. The idea was to allow the user to open with the internal editor when no other app can handle a file (the editor is a public intent for text files). Not sure what is wrong with the new Resolver.

Patch-set: 3