Update patch set 3
Patch Set 3:

I'll make this change less drastic and make it so it injects only when there aren't any other available activities then. Because as of right now, it injects even if there are other activities.

Also, as a side note, it's best to do this by just

try {
  startActivity(intent);
} catch (PackageNotFoundException e) {
  // default to internal editor
}

This way, the default resolver can be used and a lot of code can be discarded.  Binary is such a small use-case that it should be buried under a context menu if there is a default application to open it.

Patch-set: 3