Make MimeTypes Lowercase
It is standard to have mimetypes as fully lowercase in Android, so
changing the previous additions to be fully lowercase. It appears to
have worked on the device I was testing on, but other devices were
having issues. By making them all lowercase, this will make it so
it is compatible between both sets of devices.

Change-Id: I63764c0d9ec129008868ca5aa2176187834b2bc0
Issue-Id: CYNGNOS-1153