Update patch set 1
Patch Set 1:

Getting IOException while moving secure file/folder from source to destination.As per our understanding,if the file is secure file/folder while moving it getting this Exception.Now we have implemented the logic to delete the file/folder from secure storage if the file/folder is already exists then we delete the source otherwise we won't delete the file/folder.

We have used the api FileHelper.deleteFileOrFolder() to delete files/folders in secure storage.We tested here and it is working fine.

Patch-set: 1