Update patch set 4
Patch Set 4:

I think that is a good direction matt.  as long as we manage the file observers properly we shouldn't have to worry about resources.  I wonder if the directory level observer will give us the metadata of the file that was changed along with the event (modified, created, deleted, etc...)

Patch-set: 4