Update patch set 4
Patch Set 4:

Yeah, I think at a high level that's what we should do. I'm not sure about observing a directory vs a file, I've never implemented a FileObserver before. I think for the purposes of the issue that this patch is addressing: the move/copy operation within FileManager overwriting another file does not update the thumbnail -- we don't need file observers. We can just use the scrollTo parameter that Jorge is referencing to update the one file that has changed. The FileObserver solution can be implemented as a new feature later.

Patch-set: 4