Update patch set 4
Patch Set 4:

Yeah, you can monitorize a folder with a fileobserver and only listen for the needed flags for delete, create and move events. But i must say that this is something that should be done a service. Otherwise, you could lose events when the app became paused. Not sure it should worth, because monitoring the filesystem is a costly operation.

Patch-set: 4