Use System Resolver
Converting to using the system resolver.  The logic has changed slightly
in the following cases:

Open -- Tries to start activity, falls back to EditorActivity on failure
Open With -- Uses Chooser and populates Editor if not already there
	     Falls back to EditorActivity on failure
Click on FSObject -- Same as Open
Sending Object -- Same as Open

Note: The Chooser will only show EXTRA_INITIAL_INTENTS if there are
other options as well, which is why we have to do the check up front
to see if we need to add it, and falling back to just using the Editor
if nothing exists to handle it.

Change-Id: I803b86cf6bfec80bbc886025be43ce800afd3be1