Update patch set 1
Patch Set 1: Code-Review-1

From what I am seeing onRestart is not being called when toggling word wrap mode, so the dialog never shows.

Furthermore, I don't think this is a good solution.  The UI freezes due to too much work on the UI thread.  Perhaps we can show a dialog... but the UI will still freeze which is a poor experience.  I think if we want to address this issue we need to change the way large files are handled.

Patch-set: 1
Label: Code-Review=-1