Fix rename of files ending with . on sdcard
If sdcard is formatted as VFAT, then file names named as "foo." will
be silently saved as "foo".  This causes us to report the wrong target
file to media scanner, which can cause an extra file to be shown in
MTP mode on Windows.  The workaround is to rescan the parent dir to
discover the correct filename to report to media scanner.

Change-Id: Ia58a341ef6190efe3dca8b13bbc081b6e2d4e9c1
Issue-Id: QRDL-983