Fix filemanager not launch
M is using toybox instead of toolbox, some commands in filemanager
relies on toolbox utilities, there are minor difference in handling.
This change removes the "-Gn" argument pass to id command for retrival
the uid/gid, tested on both M and L, there is no need to add -Gn argument,
but in L, it seems -Gn is ignored. In M it will output wihtout the numerical
values causing parsing exception during startup.

Change-Id: I6aac8e94a80671df0a247ffcaf7d69364d7c9ad2